+++ Packages/MicrophonePitchDetector/Sources/PitchRecording/PitchRecording.swift
@@ -24,7 +24,7 @@ public struct PitchRecording: Codable, Equatable {
             throw PitchRecordingError.couldNotCreateAudioPCMBuffer
         }
 
-        let tracker = PitchTracker(sampleRate: Int32(buffer.format.sampleRate))
+        let tracker = PitchTracker(sampleRate: buffer.format.sampleRate)
 
         var iteration = 0
         var pitchRecording = PitchRecording()
